Quick fix to be able to use the AKAI Midimix
The sanitazation system of midimings prevents the sending of a NULL
byte as velocity. The AKAI Midimix mixer table is using this signal to
turns the LED off.

This patch is simply removing the line for sanitazation if the AKAI_FIX
is set. This precompiler flag is set up by setup.py in the config
section.